Updated save example script (#7899)
