+++ packages/e2e-tests/README.md
@@ -14,6 +14,18 @@
 
 ### How to record examples
 
+Examples can be re-recorded using the `save-examples.ts` script. To see all supported configuration parameters run:
+
+```sh
+./save-examples.ts --help
+```
+
+For example, to re-record the _control flow_ test using _node_ you would run:
+
+```sh
+./save-examples.ts --target=node --example=control_flow
+```
+
 ### Debugging Tips
 
 #### Cannot find a locator
