+++ packages/e2e-tests/config.ts
@@ -1,3 +1,5 @@
+import { join } from "path";
+
 import { BrowserName } from "./recordPlaywright";
 
 export default {
@@ -10,16 +12,15 @@ export default {
   driverPath: process.env.RECORD_REPLAY_DRIVER,
   headless: !!process.env.RECORD_REPLAY_PLAYWRIGHT_HEADLESS,
   nodePath: process.env.RECORD_REPLAY_NODE,
-  // This API key allows access to the whole team that holds all these test
-  // recordings. If someone wanted to, they could go in and delete the workspace
-  // or recordings in it or anything. While thats not great, it's also not
-  // the end of the world. If someone does that, we can always change
-  // this code to only run in CI in the main repo and have this as a secret.
+  nodeExamplesPath: join(__dirname, "../../test/examples/node"),
+  // This API key allows access to the whole team that holds all these test recordings.
+  // If someone wanted to, they could go in and delete the workspace or recordings in it or anything.
+  // While thats not great, it's also not the end of the world.
+  // If someone does that, we can always change this code to only run in CI in the main repo and have this as a secret.
   // It's a lot easier to hardcode it for now though.
-  // trunk-ignore(gitleaks/generic-api-key)
-  replayApiKey: "rwk_7XPbO5fhz0bkhANYXtN2dkm74wNQCchXf2OxVgAerTQ",
-  updateFixtures: !!process.env.SHOULD_UPDATE_FIXTURES,
-  useExampleFile: !process.env.SHOULD_RECORD_EXAMPLES,
+  replayApiKey: process.env.API_KEY || "rwk_7XPbO5fhz0bkhANYXtN2dkm74wNQCchXf2OxVgAerTQ",
   shouldRecordTest: !process.env.DONT_RECORD_TEST,
   shouldSaveCoverageData: !!process.env.E2E_CODE_COVERAGE,
+  updateFixtures: !!process.env.SHOULD_UPDATE_FIXTURES,
+  useExampleFile: !process.env.SHOULD_RECORD_EXAMPLES,
 };
