+++ packages/e2e-tests/package.json
@@ -20,8 +20,11 @@
     "@playwright/test": "^1.25.1",
     "@recordreplay/playwright": "^1.18.3",
     "@replayio/playwright": "0.2.27",
+    "cli-spinners": "^2.7.0",
+    "log-update": "^4",
     "playwright": "^1.25.1",
-    "ts-node": "^10.7.0"
+    "ts-node": "^10.7.0",
+    "yargs": "^17.6.0"
   },
   "dependencies": {
     "chalk": "^4"
